Refactored Attributes
