+++ src/Nox.Solution/Schema/Attributes/DescriptionAttribute.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace Nox.Solution.Schema;
+
+[AttributeUsage(AttributeTargets.Property | AttributeTargets.Field |
+   AttributeTargets.Enum | AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Interface,
+   AllowMultiple = true)]
+public class DescriptionAttribute : Attribute
+{
+    public string Description { get; private set; }
+
+    public DescriptionAttribute(string description)
+    {
+        Description = description;
+    }
+}
