+++ src/Nox.Solution/Schema/Attributes/GenerateJsonSchema.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Nox.Solution.Schema;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class GenerateJsonSchema : Attribute
+{
+    public string? SchemaName { get; private set; }
+
+    public GenerateJsonSchema(string? schemaName = null)
+    {
+        SchemaName = schemaName;
+    }
+}
