+++ src/Nox.Solution/Schema/Attributes/IgnoreAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Nox.Solution.Schema;
+
+[AttributeUsage(AttributeTargets.Property | AttributeTargets.Field, AllowMultiple = true)]
+public class IgnoreAttribute : Attribute
+{
+    public IgnoreAttribute()
+    {
+    }
+}
