+++ src/Nox.Solution/Schema/Attributes/PatternAttribute.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace Nox.Solution.Schema;
+
+[AttributeUsage(AttributeTargets.Property | AttributeTargets.Field |
+   AttributeTargets.Enum | AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Interface,
+   AllowMultiple = true)]
+public class PatternAttribute : Attribute
+{
+    public string Value { get; private set; }
+    public PatternAttribute(string value)
+    {
+        Value = value;
+    }
+}
+
