+++ src/Nox.Solution/Schema/Attributes/RequiredAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Nox.Solution.Schema;
+
+[AttributeUsage(AttributeTargets.Property | AttributeTargets.Field, AllowMultiple = true)]
+public class RequiredAttribute : Attribute
+{
+    public RequiredAttribute()
+    {
+    }
+}
